Follow-up comments on PR
